clean backend,ended  frontend and tested car,company service and controller

+++ GateFlowBackend/src/main/java/com/gateflow/GateFlow/service/CompanyService.java
@@ -1,11 +1,12 @@
 package com.gateflow.GateFlow.service;
 
+import com.gateflow.GateFlow.dto.CompanyCreateDto;
 import com.gateflow.GateFlow.dto.CompanyDto;
 import com.gateflow.GateFlow.model.Company;
 import com.gateflow.GateFlow.model.Visit;
 import com.gateflow.GateFlow.repository.CompanyRepository;
 import com.gateflow.GateFlow.repository.VisitRepository;
-import jakarta.transaction.Transactional;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,50 +14,52 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
+@Transactional(readOnly = true)
 public class CompanyService {
     private final CompanyRepository companyRepository;
-    private final VisitRepository visitRepository;
+
 @Autowired
-    public CompanyService(CompanyRepository companyRepository, VisitRepository visitRepository) {
+    public CompanyService(CompanyRepository companyRepository) {
         this.companyRepository = companyRepository;
-    this.visitRepository = visitRepository;
+
+}
+public List<CompanyDto>  findAll(){
+    return companyRepository.findAll().stream()
+            .map(c -> new CompanyDto(c.getId(),c.getName()))
+            .toList();
+}
+public CompanyDto findById(Long id){
+    return companyRepository.findById(id)
+            .map(c -> new CompanyDto(c.getId(),c.getName()))
+            .orElseThrow(() -> new RuntimeException("Nie znaleziono firmy o ID: " + id));
+
 }
 @Transactional
-    public Company addCompany(CompanyDto dto){
-    companyRepository.findByNameIgnoreCase(dto.getName())
+    public CompanyDto addCompany(CompanyCreateDto dto){
+    companyRepository.findByNameIgnoreCase(dto.name())
             .ifPresent(existing -> {
-                throw new RuntimeException("Firma o nazwie " + dto.getName() + " istnieje");
+                throw new RuntimeException("Firma o nazwie " + dto.name() + " istnieje");
             });
-    Company newCompany = new Company();
-    newCompany.setName(dto.getName());
-return companyRepository.save(newCompany);
+    Company company = new Company();
+    company.setName(dto.name());
+    Company saved = companyRepository.save(company);
+return new CompanyDto(saved.getId(),saved.getName() );
     }
-    public List<Company> findAllCompany(){
-    return companyRepository.findAll();
-    }
-    public Optional<Company> findById(Long id){
-    return companyRepository.findById(id);
-    }
-    public void deleteCompanyById(Long id){
+@Transactional
+    public void deleteCompany(Long id){
+     if (!companyRepository.existsById(id)){
+         throw new RuntimeException("Nie znaleziono firmy o ID: " + id);
+     }
      companyRepository.deleteById(id);
     }
-    public Optional<Company> findByCompanyName(String name){
-    return  companyRepository.findByNameIgnoreCase(name);
-    }
+
     @Transactional
-    public Company updatedCompany(Long id,Company request){
-    return companyRepository.findById(id)
-            .map(existing -> {
-                if (request.getName() != null) existing.setName(request.getName());
-                return companyRepository.save(existing);
-            })
-            .orElseThrow(() -> new RuntimeException("Nie znaleziono firmy o ID: " + id) );
+    public CompanyDto updateCompany(Long id, CompanyCreateDto dto){
+   Company company = companyRepository.findById(id)
+           .orElseThrow(() -> new RuntimeException("Nie znaleziono firmy o ID: " + id));
+   company.setName(dto.name());
+   return new CompanyDto(company.getId(),company.getName());
     }
-    public boolean existsById(Long id){
-    return companyRepository.existsById(id);
-    }
-    public List<Visit> findVisitsByCompanyId(Long id){
-    return visitRepository.findByCompanyId(id);
-    }
+
 
 }
